fixes #95 - refactor huge assertion class
