+++ assertpy/__init__.py
@@ -1,2 +1,3 @@
 from __future__ import absolute_import
-from .assertpy import assert_that, assert_warn, soft_assertions, contents_of, fail, soft_fail, __version__
+from .assertpy import assert_that, assert_warn, soft_assertions, fail, soft_fail, add_extension, __version__
+from .file import contents_of
