+++ assertpy/date.py
@@ -1,0 +1,81 @@
+# Copyright (c) 2015-2019, Activision Publishing, Inc.
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without modification,
+# are permitted provided that the following conditions are met:
+#
+# 1. Redistributions of source code must retain the above copyright notice, this
+# list of conditions and the following disclaimer.
+#
+# 2. Redistributions in binary form must reproduce the above copyright notice,
+# this list of conditions and the following disclaimer in the documentation
+# and/or other materials provided with the distribution.
+#
+# 3. Neither the name of the copyright holder nor the names of its contributors
+# may be used to endorse or promote products derived from this software without
+# specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+# ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
+# ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+import datetime
+
+
+class DateMixin(object):
+    """Date and time assertions mixin."""
+
+### datetime assertions ###
+    def is_before(self, other):
+        """Asserts that val is a date and is before other date."""
+        if type(self.val) is not datetime.datetime:
+            raise TypeError('val must be datetime, but was type <%s>' % type(self.val).__name__)
+        if type(other) is not datetime.datetime:
+            raise TypeError('given arg must be datetime, but was type <%s>' % type(other).__name__)
+        if self.val >= other:
+            self._err('Expected <%s> to be before <%s>, but was not.' % (self.val.strftime('%Y-%m-%d %H:%M:%S'), other.strftime('%Y-%m-%d %H:%M:%S')))
+        return self
+
+    def is_after(self, other):
+        """Asserts that val is a date and is after other date."""
+        if type(self.val) is not datetime.datetime:
+            raise TypeError('val must be datetime, but was type <%s>' % type(self.val).__name__)
+        if type(other) is not datetime.datetime:
+            raise TypeError('given arg must be datetime, but was type <%s>' % type(other).__name__)
+        if self.val <= other:
+            self._err('Expected <%s> to be after <%s>, but was not.' % (self.val.strftime('%Y-%m-%d %H:%M:%S'), other.strftime('%Y-%m-%d %H:%M:%S')))
+        return self
+
+    def is_equal_to_ignoring_milliseconds(self, other):
+        if type(self.val) is not datetime.datetime:
+            raise TypeError('val must be datetime, but was type <%s>' % type(self.val).__name__)
+        if type(other) is not datetime.datetime:
+            raise TypeError('given arg must be datetime, but was type <%s>' % type(other).__name__)
+        if self.val.date() != other.date() or self.val.hour != other.hour or self.val.minute != other.minute or self.val.second != other.second:
+            self._err('Expected <%s> to be equal to <%s>, but was not.' % (self.val.strftime('%Y-%m-%d %H:%M:%S'), other.strftime('%Y-%m-%d %H:%M:%S')))
+        return self
+
+    def is_equal_to_ignoring_seconds(self, other):
+        if type(self.val) is not datetime.datetime:
+            raise TypeError('val must be datetime, but was type <%s>' % type(self.val).__name__)
+        if type(other) is not datetime.datetime:
+            raise TypeError('given arg must be datetime, but was type <%s>' % type(other).__name__)
+        if self.val.date() != other.date() or self.val.hour != other.hour or self.val.minute != other.minute:
+            self._err('Expected <%s> to be equal to <%s>, but was not.' % (self.val.strftime('%Y-%m-%d %H:%M'), other.strftime('%Y-%m-%d %H:%M')))
+        return self
+
+    def is_equal_to_ignoring_time(self, other):
+        if type(self.val) is not datetime.datetime:
+            raise TypeError('val must be datetime, but was type <%s>' % type(self.val).__name__)
+        if type(other) is not datetime.datetime:
+            raise TypeError('given arg must be datetime, but was type <%s>' % type(other).__name__)
+        if self.val.date() != other.date():
+            self._err('Expected <%s> to be equal to <%s>, but was not.' % (self.val.strftime('%Y-%m-%d'), other.strftime('%Y-%m-%d')))
+        return self
